feat: Top,middle category 삭제 기능

+++ src/main/java/com/multitab/category/cate/infrastructure/MiddleCategoryRepository.java
@@ -7,11 +7,15 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface MiddleCategoryRepository extends JpaRepository<MiddleCategory, Integer> {
+public interface MiddleCategoryRepository extends JpaRepository<MiddleCategory, Long> {
 
     List<MiddleCategory> findByTopCategoryCategoryCode(String topCategoryCode);
+
     List<MiddleCategory> findByTopCategoryCategoryName(String topCategoryCode);
+
     Optional<MiddleCategory> findByCategoryCode(String categoryCode);
+
     boolean existsByCategoryCode(String categoryCode);
+
     boolean existsByCategoryName(String categoryName);
 }
